+++ flutter/lib/desktop/pages/remote_page.dart
@@ -328,8 +328,7 @@ class _RemotePageState extends State<RemotePage>
           Stack(
             children: [
               _ffi.ffiModel.pi.isSet.isTrue &&
-                      _ffi.ffiModel.waitForFirstImage.isTrue &&
-                      !_ffi.ffiModel.isPeerAndroid
+                      _ffi.ffiModel.waitForFirstImage.isTrue
                   ?  /*Obx(() => Offstage(
                       offstage:
                           _ffi.dialogManager.mobileActionsOverlayVisible.isFalse,